+++ crates/voicevox_core/src/core/voice_model.rs
@@ -35,8 +35,10 @@ use super::{
         },
     },
     manifest::{
-        Manifest, ManifestDomain, ManifestDomains, ModelFile, ModelFileType, StyleIdToInnerVoiceId,
+        FormatVersion, Manifest, ManifestDomain, ManifestDomains, ModelFile, ModelFileType,
+        StyleIdToInnerVoiceId,
     },
+    metas::{VoiceModelMetaSchemaV1, VoiceModelMetaSchemaV2},
 };
 
 pub(super) type ModelBytesWithInnerVoiceIdsByDomain = inference_domain_map_values!(
@@ -455,13 +457,30 @@ impl VoiceModelHeader {
             source: Some(source),
         };
 
-        let metas = serde_json::from_slice::<VoiceModelMeta>(metas).map_err(|source| {
-            error(
-                LoadModelErrorKind::InvalidModelFormat,
-                anyhow::Error::from(source)
-                    .context(format!("{}が不正です", manifest.metas_filename())),
-            )
-        })?;
+        let metas = match manifest.vvm_format_version() {
+            FormatVersion::V1 => {
+                let data =
+                    serde_json::from_slice::<VoiceModelMetaSchemaV1>(metas).map_err(|source| {
+                        error(
+                            LoadModelErrorKind::InvalidModelFormat,
+                            anyhow::Error::from(source)
+                                .context(format!("{}が不正です", manifest.metas_filename())),
+                        )
+                    })?;
+                VoiceModelMeta::from_iter(data.into_iter().map(Into::into))
+            }
+            FormatVersion::V2 => {
+                let data =
+                    serde_json::from_slice::<VoiceModelMetaSchemaV2>(metas).map_err(|source| {
+                        error(
+                            LoadModelErrorKind::InvalidModelFormat,
+                            anyhow::Error::from(source)
+                                .context(format!("{}が不正です", manifest.metas_filename())),
+                        )
+                    })?;
+                VoiceModelMeta::from_iter(data.into_iter().map(Into::into))
+            }
+        };
 
         manifest
             .domains()